added square adjustments, prune in root, unmove paradigm

+++ C++/strategy.h
@@ -6,6 +6,6 @@ const int search_depth = 5;
 
 Move best_move(Board &position);
 Move best_move_minimax(Board position);
-float negamax(Board &position, int depth, float alpha, float beta, bool max_player);
-float minimax(Board position, int depth, float alpha, float beta, bool max_player);
+short negamax(Board &position, int depth, short alpha, short beta, bool max_player);
+short minimax(Board position, int depth, short alpha, short beta, bool max_player);
 #endif
